requirements.txt
beautifulsoup4==4.11.1
docx==0.2.4
matplotlib==3.5.1
progressbar33==2.4
python_docx==0.8.11
requests==2.27.1
